refactor interface
