+++ tabulous/widgets/_component/__init__.py
@@ -1,0 +1,8 @@
+from ._base import Component, TableComponent, ViewerComponent
+from ._header import VerticalHeaderInterface, HorizontalHeaderInterface
+from ._ranges import SelectionRanges, HighlightRanges, SelectedData
+from ._cell import CellInterface
+from ._cellref import CellReferenceInterface
+from ._plot import PlotInterface
+from ._spreadsheet import ColumnDtypeInterface
+from ._proxy import ProxyInterface
